% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HamONdest-package.R
\docType{data}
\name{Care_Facilities}
\alias{Care_Facilities}
\title{Care_Facilities}
\format{
A simple feature class (sf) point object containing 668 rows and 11 variables; each row represents a health/community care facility/service and thus addresses may be repeated if the location hosts multiple facilities/services.
\describe{
\item{Name}{Name of facilities as specifed by the City of Hamilton or by Ontario Ministry of Health .}
\item{Service_Type}{The type of service offered as described by the the Ontario Ministry of Health service provider locations feature layer documentation.}
\item{Service_Desc}{The service as described by the the Ontario Ministry of Health service provider locations feature layer documentation.}
\item{Address}{The address as supplied by the City of Hamilton or the Ontario Ministry of Health service provider locations feature layer documentation.}
\item{Address_2}{Additional address details (unit number, floor, etc.).}
\item{Postalcode}{Postal code.}
\item{Address_Desc}{The address of the service may be either the "Pratice Location", "Administrative Office", "Administration and Practice Location", or blank. This field is specified by the the Ontario Ministry of Health service provider locations feature layer documentation.}
\item{Community}{The community in which the service address is located in; either "Hamilton", "Burlington", "Grimbsy", or "NA".}
\item{Type}{Type of facility, either "Community or Health Support Service", "Primary Health Service", "Fee-Based Independent Health Facility", "Hospital", "Laboratory - Specimen Collection Centre", or "Long-Term Care Home", "Pharmacy", or "Residential Care Facility" .}
\item{Supply}{The number of 'supply' offered by the facility; for "Residential Care Facility" type this is the capacity range available at the location as specified by the City of Hamilton Open Data layer; all other types are unassigned (for now).}
\item{geometry}{The sfc point geometry (location) of each facility/service.}
}
}
\source{
"Licensed residential care facilities" from \href{https://open.hamilton.ca/datasets/e37b13a002544e359350c31b2d48ee47_10/about}{Hamilton Open Data} accessed November 10 2021; "Ministry of Health service provider locations" from \href{https://geohub.lio.gov.on.ca/datasets/lio::ministry-of-health-service-provider-locations/about}{Ontario GeoHub} accessed November 14 2021.
}
\usage{
data(Care_Facilities)
}
\description{
This object a few types of health and community care facilities/services (Hospitals, Residential Care Facilities, Long-Term Care Homes, Community or Health Support Services, Independent Health Facilities, Pharmacies, and Primary Health Services) represented as spatial points.
}
\keyword{Hamilton}
\keyword{care}
\keyword{clinic}
\keyword{health}
\keyword{hospital}
\keyword{residential}
