% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HamONdest-package.R
\docType{data}
\name{Ham_CityBound}
\alias{Ham_CityBound}
\title{Ham_CityBound}
\format{
A simple feature class (sfc) multipolygon object containing 1 row.
}
\source{
"City boundary" from \href{https://open.hamilton.ca/datasets/489f656072f44f34b07bf53efa0d8e4f_13/about}{Hamilton Open Data} accessed November 8 2021.
}
\usage{
data(Ham_CityBound)
}
\description{
This object defines the spatial boundary of the City of Hamilton.
}
\keyword{Hamilton}
\keyword{boundary}
\keyword{city}
