% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HamONdest-package.R
\docType{data}
\name{Ham_WardBound}
\alias{Ham_WardBound}
\title{Ham_WardBound}
\format{
A simple feature class (sf) polygon object containing 15 rows and 2 variables.
\describe{
\item{WARD}{A unique numeric ID for each ward in Hamilton. Each ward serves as an electoral boundary and is assigned a voted-in city councillor.}
\item{geometry}{The sfc polygon geometry of each ward boundary}
}
}
\source{
"Ward Boundaries" from \href{https://open.hamilton.ca/datasets/8b0b1f2bf8bb4e1da3a1bf567b17b77f_7/about}{Hamilton Open Data} accessed November 8 2021.
}
\usage{
data(Ham_WardBound)
}
\description{
This object defines the spatial boundary of the City of Hamilton.
}
\keyword{Hamilton}
\keyword{boundaries}
\keyword{ward}
