% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HamONdest-package.R
\docType{package}
\name{HamONdest-package}
\alias{HamONdest}
\alias{HamONdest-package}
\title{HamONdest: Potential destinations in Hamilton, Ontario, Canada: for transportation planning}
\description{
The goal of HamONdest is to compile all potential destinations (Schools, parks, municipally-owned land, libraries, etc.) in Hamilton, Ontario, Canada all in one data-package to facilitate the calculation of accessibility and mobility measures in the transportation planning context.

The goal of HamONdest is to compile all potential destinations (Schools, parks, municipally-owned land, libraries, etc.) in Hamilton, Ontario, Canada all in one data-package to facilitate the calculation of accessibility and mobility measures in the transportation planning context.
}
\references{
\url{https://github.com/soukhova/HamONdest}
}
\author{
\strong{Maintainer}: Anastasia Soukhov \email{soukhoa@mcmaster.ca}


Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
\keyword{internal}
