% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HamONdest-package.R
\docType{data}
\name{Schools_Catchments_201516_201011}
\alias{Schools_Catchments_201516_201011}
\title{Schools_Catchments_201516_201011}
\format{
A simple feature (sf) polygon object containing 305 rows and 5 variables :
\describe{
\item{CID}{A unique integer ID fpr each catchment.}
\item{geometry}{The sf polygon geometry}
\item{Year}{A character tag identifying what year of the catchment (either "2011 and 2016", "2011" or "2016" where 2011 and 2016 is shorthand for 2010-2011 and 2015-2016 respectively).}
\item{Level}{A characer tag identifying the level of the catchment (either "High" or "Elementary"). Elementary indicates instruction of grades JK through 8 (approx. ages 5 to 13) and Secondary indicates instruction of grades 9 through 12 (approx. ages 14 to 18)}
\item{System}{A character tag identifying the system of the catchment (either "Public" or "Catholic")}
}
}
\source{
Catchment ID, geometry, year, level, and system are sourced from spatial polygons provided directly by the Hamilton Wentworth District School Board (HWDSB) and the Hamilton Wentworth Catholic District School Board (HWCDSB) for the respective academic years (2010-2011 and 2015-2016).
}
\usage{
data(Schools_Catchments_201516_201011)
}
\description{
A simple feature object containing all public and public catholic schools catchments (elementary and secondary) in Hamilton, Ontario, Canada for the 2015-2016 and 2010-2011 academic years.
This object should be used alongside 'Schools_201516_201011'.
}
\examples{
 data(Schools_201516_201011)
 CID <- Schools_Catchments_201516_201011$CID
}
\keyword{catchment}
\keyword{school}
\keyword{student}
